+++ packages/eas-cli/src/credentials/android/actions/AssignGoogleServiceAccountKey.ts
@@ -3,14 +3,14 @@ import {
   GoogleServiceAccountKeyFragment,
 } from '../../../graphql/generated';
 import Log from '../../../log';
-import { Context } from '../../context';
+import { CredentialsContext } from '../../context';
 import { AppLookupParams } from '../api/GraphqlClient';
 
 export class AssignGoogleServiceAccountKey {
   constructor(private app: AppLookupParams) {}
 
   public async runAsync(
-    ctx: Context,
+    ctx: CredentialsContext,
     googleServiceAccountKey: GoogleServiceAccountKeyFragment
   ): Promise<CommonAndroidAppCredentialsFragment> {
     const appCredentials =