+++ packages/eas-cli/src/credentials/ios/actions/CreatePushKey.ts
@@ -1,13 +1,13 @@
 import { ApplePushKeyFragment } from '../../../graphql/generated';
 import Log from '../../../log';
 import { Account } from '../../../user/Account';
-import { Context } from '../../context';
+import { CredentialsContext } from '../../context';
 import { provideOrGeneratePushKeyAsync } from './PushKeyUtils';
 
 export class CreatePushKey {
   constructor(private account: Account) {}
 
-  public async runAsync(ctx: Context): Promise<ApplePushKeyFragment> {
+  public async runAsync(ctx: CredentialsContext): Promise<ApplePushKeyFragment> {
     if (ctx.nonInteractive) {
       throw new Error(`A new push key cannot be created in non-interactive mode.`);
     }