+++ packages/eas-cli/src/credentials/manager/HelperActions.ts
@@ -1,5 +1,10 @@
 import Log from '../../log';
 import { pressAnyKeyToContinueAsync } from '../../prompts';
+import { CredentialsContext } from '../context';
+
+export interface Action<T = void> {
+  runAsync(ctx: CredentialsContext): Promise<T>;
+}
 
 export class PressAnyKeyToContinue {
   public async runAsync(): Promise<void> {